Remove issuestats and userstats commands.
